pyproject.toml
[project]
name = "pyphony"
version = "0.1.0"
description = "Python implementation of the Symphony Service Specification"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "httpx>=0.27",
    "structlog>=24.0",
    "watchfiles>=0.21",
    "starlette>=0.37",
    "uvicorn>=0.29",
    "claude-agent-sdk>=0.1.48",
    "python-dotenv>=1.0",
]

[project.scripts]
pyphony = "pyphony.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pyphony"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
]
